int -> uint_t

+++ Onboard-SDK/project_folder/transceiver_search/transceiver_search.hpp
@@ -48,7 +48,7 @@ bool runWaypointMission(DJI::OSDK::Vehicle* vehicle, uint8_t numWaypoints, int r
 void setWaypointDefaults(DJI::OSDK::WayPointSettings* wp);
 void setWaypointInitDefaults(DJI::OSDK::WayPointInitSettings* fdata);
 
-std::vector<DJI::OSDK::WayPointSettings> createWaypoints(DJI::OSDK::Vehicle* vehicle, int numWaypoints, float64_t latM,
+std::vector<DJI::OSDK::WayPointSettings> createWaypoints(DJI::OSDK::Vehicle* vehicle, uint8_t numWaypoints, float64_t latM,
                                                          float64_t lonM, DJI::OSDK::float32_t start_alt);
 
 std::vector<DJI::OSDK::WayPointSettings> generateWaypoints(DJI::OSDK::WayPointSettings* start_data, int num_wp,
